refactor(visor): split rpc.go by topic to mirror api_*.go layout
rpc.go was 2185 lines containing 186 RPC handler methods across every
topic the visor exposes. The package already has topic-split companion
files (api_apps.go, api_routing.go, api_dmsg.go, ...) but their RPC
handler counterparts all lived in one file, making PR diffs noisy and
navigation hard.

Split rpc.go into 13 topic files mirroring the existing api_*.go
groupings:

  rpc_apps.go       (27 handlers, 212 lines)  app start/stop/config
  rpc_ar.go         (1)            address-resolver helpers
  rpc_dht.go        (6)            DHT operations
  rpc_dmsg.go       (13)           DMSG probes / sessions / bandwidth
  rpc_hypervisors.go(28)           HV* proxy methods
  rpc_network.go    (6)            skynet/forwarded TCP ports
  rpc_ping.go       (13)           ping / dmsg ping
  rpc_proxies.go    (9)            embedded proxies, raw TCP forwards
  rpc_routing.go    (16)           rules, route groups, mux routes
  rpc_services.go   (4)            VPN/proxy server discovery
  rpc_tps.go        (7)            transport-setup-node management
  rpc_transport.go  (22)           transport CRUD, persistence, public
  rpc_visor.go      (26)           lifecycle, summary, health, uptime,
                                    rewards, reload, shutdown, ui

rpc.go shrinks to 385 lines: package + imports + RPC struct +
newRPCServer + the ~47 input/output type definitions shared across
files. No behavior change.
